requirements.txt
streamlit>=1.34
pandas>=2.1
requests>=2.31
tldextract>=5.1
google-genai>=0.6.0
openai>=1.30.0
anthropic>=0.30.0
python-dateutil>=2.9
